更新API接口至https

+++ api/app/Http/Controllers/Api/V1/UserController.php
@@ -1,31 +1,32 @@
 <?php
-/**
- * MIT License
- *
- * Copyright (c) 2017 Mrdai
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-namespace App\Http\Controllers\Api\V1;
+	/**
+	 * MIT License
+	 *
+	 * Copyright (c) 2017 Mrdai
+	 *
+	 * Permission is hereby granted, free of charge, to any person obtaining a copy
+	 * of this software and associated documentation files (the "Software"), to deal
+	 * in the Software without restriction, including without limitation the rights
+	 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+	 * copies of the Software, and to permit persons to whom the Software is
+	 * furnished to do so, subject to the following conditions:
+	 *
+	 * The above copyright notice and this permission notice shall be included in all
+	 * copies or substantial portions of the Software.
+	 *
+	 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+	 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+	 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+	 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+	 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+	 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+	 * SOFTWARE.
+	 */
+	
+	namespace App\Http\Controllers\Api\V1;
 	
 	use Illuminate\Http\Request;
+	use Illuminate\Validation\ValidationException;
 	use App\Http\Controllers\Api\ApiController;
 	use App\Http\Resources\UserCollection;
 	use App\Http\Resources\UserResource;
@@ -91,6 +92,12 @@ namespace App\Http\Controllers\Api\V1;
 		 */
 		public function update( \App\Http\Requests\UpdateUserRequest $request, $id ) {
 			$user = User::find( $id );
+			if ( ! $user ) {
+				throw ValidationException::withMessages( [
+					"user" => [ "400005" ],
+				] )->status( 400 );
+			}
+			
 			$request->has( 'username' ) ? $user->username = $request->get( 'username' ) : null;
 			$request->has( 'password' ) ? $user->password = bcrypt( $request->get( 'password' ) ) : null;
 			$request->has( 'email' ) ? $user->email = $request->get( 'email' ) : null;
@@ -111,6 +118,13 @@ namespace App\Http\Controllers\Api\V1;
 		 */
 		public function destroy( $id ) {
 			$user = User::find( $id );
+			
+			if ( ! $user ) {
+				throw ValidationException::withMessages( [
+					"user" => [ "400005" ],
+				] )->status( 400 );
+			}
+			
 			$user->resources()->each( function ( $item, $key ) {
 				Resource::destroy( $item->id );
 			} );
